Move code for Example base classes to Examples/Common/
Had to also update most Examples to stop using relative paths for their includes.
