+++ Examples/AtfCCSD/AtfCCSD.cpp
@@ -1,4 +1,4 @@
-#include "../ExampleBase.h"
+#include "ExampleBase.h"
 #include <memory>
 
 using namespace std;
