+++ Examples/Common/CliComponent.cpp
