+++ Examples/Common/CliComponent.h
