+++ Examples/Common/ExampleBase.cpp
