+++ Examples/Common/ExampleBase.h
