+++ Examples/Common/ExampleReferenceKernel.cpp
