+++ Examples/Common/ExampleReferenceKernel.h
