+++ Examples/Common/RunStats.hpp
