+++ Examples/CoulombSum2d/CoulombSum2d.cpp
@@ -1,4 +1,4 @@
-#include "../ExampleReferenceKernel.h"
+#include "ExampleReferenceKernel.h"
 #include <memory>
 
 using namespace std;
