+++ premake5.lua
@@ -404,7 +404,7 @@ function addExampleProject(name, kernelExt, apiDefine, useRefVersions, shouldEna
     project(projectName)
         kind "ConsoleApp"
         files {table.unpack(cppFiles)}
-        includedirs {"Source", "Examples"}
+        includedirs {"Source", "Examples/Common"}
         defines {apiDefine}
         links {"ktt", exLib}
         if shouldEnableOpenMP then
@@ -698,7 +698,7 @@ project "ExamplesLibCuda"
     kind "StaticLib"
     files
     {
-        "Examples/*.cpp"
+        "Examples/Common/*.cpp"
     }
     includedirs {"Source"}
     defines {"KTT_CUDA_EXAMPLE"}
@@ -707,7 +707,7 @@ project "ExamplesLibOpenCl"
     kind "StaticLib"
     files
     {
-        "Examples/*.cpp"
+        "Examples/Common/*.cpp"
     }
     includedirs {"Source"}
     defines {"KTT_OPENCL_EXAMPLE"}
@@ -716,7 +716,7 @@ project "ExamplesLibCpp"
     kind "StaticLib"
     files
     {
-        "Examples/*.cpp"
+        "Examples/Common/*.cpp"
     }
     includedirs {"Source"}
     defines {"KTT_CPP_EXAMPLE"}
